+++ src/lib/components/destinations/places/Fractal.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
-	import DestinationCard from '$lib/components/DestinationCard.svelte';
+	import DestinationCard from '../DestinationCard.svelte';
 	import type { DistancesResult, GoalStatus } from '$lib/types';
 
-	let { transit }: DistancesResult['church'] = $props();
+	let { transit }: DistancesResult['fractal'] = $props();
 
 	const maxMinutes = 30;
 
@@ -12,7 +12,7 @@
 </script>
 
 <DestinationCard
-	title="Church ⛪️ "
+	title="Fractal 🪩"
 	goal="Ideally within {maxMinutes} minutes (low priority)"
 	routes={[`🚇 ${transit.timeMinutes} minutes on the ${transit.summary}`]}
 	goalStatus={goalStatus()}
